ajout et affichage vehicule

+++ src/main/java/com/accenture/applicationlocationvehicule/repository/BikeDao.java
@@ -3,5 +3,5 @@ package com.accenture.applicationlocationvehicule.repository;
 import com.accenture.applicationlocationvehicule.repository.entity.Bike;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BikeDao extends JpaRepository<Bike , Integer> {
+public interface BikeDao extends JpaRepository<Bike , Long> {
 }
