+++ src/main/java/com/accenture/applicationlocationvehicule/repository/CarDao.java
@@ -3,5 +3,5 @@ package com.accenture.applicationlocationvehicule.repository;
 import com.accenture.applicationlocationvehicule.repository.entity.Car;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CarDao extends JpaRepository<Car, Integer> {
+public interface CarDao extends JpaRepository<Car, Long> {
 }
