+++ src/main/java/com/accenture/applicationlocationvehicule/repository/MotorcycleDao.java
@@ -3,5 +3,5 @@ package com.accenture.applicationlocationvehicule.repository;
 import com.accenture.applicationlocationvehicule.repository.entity.Motorcycle;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MotorcycleDao extends JpaRepository<Motorcycle , Integer> {
+public interface MotorcycleDao extends JpaRepository<Motorcycle , Long> {
 }
