+++ src/main/java/com/accenture/applicationlocationvehicule/repository/MotorhomeDao.java
@@ -3,5 +3,5 @@ package com.accenture.applicationlocationvehicule.repository;
 import com.accenture.applicationlocationvehicule.repository.entity.Motorhome;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MotorhomeDao extends JpaRepository<Motorhome, Integer> {
+public interface MotorhomeDao extends JpaRepository<Motorhome, Long> {
 }
