+++ src/main/java/com/accenture/applicationlocationvehicule/repository/UtilityVehicleDao.java
@@ -3,5 +3,5 @@ package com.accenture.applicationlocationvehicule.repository;
 import com.accenture.applicationlocationvehicule.repository.entity.UtilityVehicle;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UtilityVehicleDao extends JpaRepository<UtilityVehicle , Integer> {
+public interface UtilityVehicleDao extends JpaRepository<UtilityVehicle , Long> {
 }
